DB migrations using Flyway, split Main class
